requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pymysql==1.1.1
python-dotenv==1.0.1
pydantic[email]==2.10.3
cryptography==44.0.0